// src/app/model/companyClient.ts
import { User } from "./user";

export class CompanyClient {
    idClientC : number;
    brand:string;
    activityArea: string ;
    creationDate: Date ;
    employeesNb: number ;
    capital: number ;
    nbDeclaredSinistersC: number ;
    archived: boolean ;
    sbuscriptionDate: Date ;
    categoryC: string ;
    logClientC:User;
    ScoreC :number;
}